Configure Superset to use OpenShift OAuth (#510)
* Break out superset_config.py into a standalone file

Managing this file as a python file in the repo rather than as an inine
script in a secret definition makes it much easier to write (since we
can use our pythoneditor of choice to do so).

* Fix flake8 errors in superset_config.py

* Allow for customizing superset_config

This adds support for adding additional config to superset_config.py by
way of an extra script. If the path to the script is defined, it will be
dynamically executed, allowing for ODH users to customize their
superset deployment without overriding the base functionality in ODH.

* Switch the Superset route to use HTTPS

* Add more sane defaults to the superset_config

* Enable oauth for superset

* Add instructions for customizing the superset_config.py script

* Move superset_config.py to a configmap

This file contained no sensitive information, so a ConfigMap is more
appropriate than a secret for maintaining it.

* Update expected pod count for superset test

With this PR I changed the kustomize to add an `app=superset` label to
all of the Superset objects. This means that we now have the label on
both the superset and superset db pods. We need to change the expected
pod count from 1 to 2 in the test to account for this.