Fix certification verification issue for superset oauth (#517)
* Fix certification verification issue for superset oauth

In the code that was added for Superset OAuth, we specify the path to
the certificate bundle to be used for verifying requests to the local
openshift cluster's oauth endpoint.

The cert that was used is the openshift cluster certificate. It turns
out that in cases when a globally trusted certificate is used (rather
than a cluster self-signed cert), this openshift cluster certificate
cannot be used to validate the oauth endpoint. Instead, in these cases
we must use the default system certificate bundle in /etc/ssl/certs.
Unfortunately, this certificate bundle does not work on clusters that
generate a self-signed cluster cert.

This change combines the default global cert bundle with the openshift
cluster cert bundle and writes them together into a new file. This
resulting cert bundle file is then used by superset to perform oauth and
works on both types of clusters.

* Handle case when oauth user full name does not exist

In some identity providers, the user's full name field may not exist.
This case broke superset oauth. This change adds handling for this. When
this is the case, the username in the superset database will be
populated, but none of the first or last name fields will be. Superset
functions just fine like this.