Update Superset OAuth role code to account for nonexistent Role (#539)
The superset OAuth code loops through a user's groups as returned by the
OAuth server (OpenShift) and attempts to find a corresponding Superset
role. If no such match exists, Superset attempts to save a null entry in
the database. This doesn't yield any user-facing errors, but prints an
ugly error message in the pod logs.

This code changes the logic for looping through the groups and adds a
check for null and duplicate values before saving the user's roles.